move non-state impact functions into separate file for code clarity
- move methods into formMethods
- move render buttons (editable and non-editable mode buttons) into FormBtns
